// src/models/utils/ISeoMetadata.ts
/**
 * Provides the functionalities of SEO metadata.
 */
export interface ISeoMetadata {
  /**
   * Description.
   */
  description?: string

  /**
   * Image alt text.
   */
  imageAlt?: string

  /**
   * Image URL.
   */
  image?: string

  /**
   * Title.
   */
  title?: string

  /**
   * URL.
   */
  url?: string
}